Move definition tests to new base class
