+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Definition/RazorDefinitionEndpointTest.cs
@@ -7,32 +7,18 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
-using Microsoft.AspNetCore.Razor.LanguageServer.Common;
 using Microsoft.AspNetCore.Razor.LanguageServer.Completion;
-using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
 using Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
-using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
-using Microsoft.AspNetCore.Razor.LanguageServer.Test;
 using Microsoft.AspNetCore.Razor.LanguageServer.Test.Common;
-using Microsoft.AspNetCore.Razor.Test.Common;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
-using Microsoft.CodeAnalysis.Razor.Workspaces;
 using Microsoft.CodeAnalysis.Razor.Workspaces.Extensions;
 using Microsoft.CodeAnalysis.Testing;
 using Microsoft.CodeAnalysis.Text;
-using Microsoft.VisualStudio.LanguageServer.Protocol;
 using Moq;
-using OmniSharp.Extensions.JsonRpc;
-using OmniSharp.Extensions.LanguageServer.Protocol.Server;
 using Xunit;
-using DefinitionResult = Microsoft.VisualStudio.LanguageServer.Protocol.SumType<
-    Microsoft.VisualStudio.LanguageServer.Protocol.VSInternalLocation,
-    Microsoft.VisualStudio.LanguageServer.Protocol.VSInternalLocation[],
-    Microsoft.VisualStudio.LanguageServer.Protocol.DocumentLink[]>;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Definition
 {
-    [UseExportProvider]
     public class RazorDefinitionEndpointTest : TagHelperServiceTestBase
     {
         private const string DefaultContent = @"@addTagHelper *, TestAssembly
@@ -426,186 +412,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Definition
             var outDoc = codeDocument;
             documentSnapshot = Mock.Of<DocumentSnapshot>(d => d.GetTextAsync() == Task.FromResult(sourceText), MockBehavior.Strict);
             Mock.Get(documentSnapshot).Setup(s => s.GetGeneratedOutputAsync()).Returns(Task.FromResult(outDoc));
-        }
-
-        [Fact]
-        public async Task Handle_SingleServer_CSharp_Method()
-        {
-            var input = """
-                <div></div>
-
-                @{
-                    var x = Ge$$tX();
-                }
-
-                @functions
-                {
-                    void [|GetX|]()
-                    {
-                    }
-                }
-                """;
-
-            await VerifyCSharpGoToDefinitionAsync(input);
-        }
-
-        [Fact]
-        public async Task Handle_SingleServer_CSharp_Local()
-        {
-            var input = """
-                <div></div>
-
-                @{
-                    var x = GetX();
-                }
-
-                @functions
-                {
-                    private string [|_name|];
-
-                    string GetX()
-                    {
-                        return _na$$me;
-                    }
-                }
-                """;
-
-            await VerifyCSharpGoToDefinitionAsync(input);
-        }
-
-        [Fact]
-        public async Task Handle_SingleServer_CSharp_MetadataReference()
-        {
-            var input = """
-                <div></div>
-
-                @functions
-                {
-                    private stri$$ng _name;
-                }
-                """;
-
-            // Arrange
-            TestFileMarkupParser.GetPosition(input, out var output, out var cursorPosition);
-
-            var codeDocument = CreateCodeDocument(output);
-            var razorFilePath = "C:/path/to/file.razor";
-
-            // Act
-            var result = await GetDefinitionResultAsync(codeDocument, razorFilePath, cursorPosition);
-
-            // Assert
-            Assert.NotNull(result.Value.Second);
-            var locations = result.Value.Second;
-            var location = Assert.Single(locations);
-            Assert.EndsWith("String.cs", location.Uri.ToString());
-            Assert.Equal(21, location.Range.Start.Line);
-        }
-
-        private async Task VerifyCSharpGoToDefinitionAsync(string input)
-        {
-            // Arrange
-            TestFileMarkupParser.GetPositionAndSpan(input, out var output, out var cursorPosition, out var expectedSpan);
-
-            var codeDocument = CreateCodeDocument(output);
-            var razorFilePath = "C:/path/to/file.razor";
-
-            // Act
-            var result = await GetDefinitionResultAsync(codeDocument, razorFilePath, cursorPosition);
-
-            // Assert
-            Assert.NotNull(result.Value.Second);
-            var locations = result.Value.Second;
-            var location = Assert.Single(locations);
-            Assert.Equal(new Uri(razorFilePath), location.Uri);
-
-            var expectedRange = expectedSpan.AsRange(codeDocument.GetSourceText());
-            Assert.Equal(expectedRange, location.Range);
-        }
-
-        private async Task<DefinitionResult?> GetDefinitionResultAsync(RazorCodeDocument codeDocument, string razorFilePath, int cursorPosition)
-        {
-            var realLanguageServerFeatureOptions = new DefaultLanguageServerFeatureOptions();
-
-            var csharpSourceText = codeDocument.GetCSharpSourceText();
-            var serverCapabilities = new ServerCapabilities()
-            {
-                DefinitionProvider = true
-            };
-            var csharpDocumentUri = new Uri(realLanguageServerFeatureOptions.GetRazorCSharpFilePath(razorFilePath));
-            var csharpServer = await CSharpTestLspServerHelpers.CreateCSharpLspServerAsync(csharpSourceText, csharpDocumentUri, serverCapabilities, razorSpanMappingService: null).ConfigureAwait(false);
-            await csharpServer.OpenDocumentAsync(csharpDocumentUri, csharpSourceText.ToString()).ConfigureAwait(false);
-
-            var documentContextFactory = new TestDocumentContextFactory(razorFilePath, codeDocument, version: 1337);
-            var languageServerFeatureOptions = Mock.Of<LanguageServerFeatureOptions>(options =>
-                options.SupportsFileManipulation == true &&
-                options.SingleServerSupport == true &&
-                options.CSharpVirtualDocumentSuffix == realLanguageServerFeatureOptions.CSharpVirtualDocumentSuffix &&
-                options.HtmlVirtualDocumentSuffix == realLanguageServerFeatureOptions.HtmlVirtualDocumentSuffix
-                , MockBehavior.Strict);
-            var languageServer = new DefinitionLanguageServer(csharpServer, csharpDocumentUri);
-            var documentMappingService = new DefaultRazorDocumentMappingService(languageServerFeatureOptions, documentContextFactory, LoggerFactory);
-            var projectSnapshotManager = Mock.Of<ProjectSnapshotManagerBase>(p => p.Projects == new[] { Mock.Of<ProjectSnapshot>(MockBehavior.Strict) }, MockBehavior.Strict);
-            var projectSnapshotManagerAccessor = new TestProjectSnapshotManagerAccessor(projectSnapshotManager);
-            var projectSnapshotManagerDispatcher = new LSPProjectSnapshotManagerDispatcher(LoggerFactory);
-            var searchEngine = new DefaultRazorComponentSearchEngine(Dispatcher, projectSnapshotManagerAccessor, LoggerFactory);
-
-            var endpoint = new RazorDefinitionEndpoint(documentContextFactory, searchEngine, documentMappingService, languageServerFeatureOptions, languageServer, TestLoggerFactory.Instance);
-
-            codeDocument.GetSourceText().GetLineAndOffset(cursorPosition, out var line, out var offset);
-            var request = new DefinitionParamsBridge
-            {
-                TextDocument = new TextDocumentIdentifier
-                {
-                    Uri = new Uri(razorFilePath)
-                },
-                Position = new Position(line, offset)
-            };
-
-            return await endpoint.Handle(request, CancellationToken.None);
-        }
-
-        private class DefinitionLanguageServer : ClientNotifierServiceBase
-        {
-            private readonly CSharpTestLspServer _csharpServer;
-            private readonly Uri _csharpDocumentUri;
-
-            public DefinitionLanguageServer(CSharpTestLspServer csharpServer, Uri csharpDocumentUri)
-            {
-                _csharpServer = csharpServer;
-                _csharpDocumentUri = csharpDocumentUri;
-            }
-
-            public override OmniSharp.Extensions.LanguageServer.Protocol.Models.InitializeParams ClientSettings { get; }
-
-            public override Task OnStarted(ILanguageServer server, CancellationToken cancellationToken)
-            {
-                return Task.CompletedTask;
-            }
-
-            public override Task<IResponseRouterReturns> SendRequestAsync(string method)
-            {
-                throw new NotImplementedException();
-            }
-
-            public async override Task<IResponseRouterReturns> SendRequestAsync<T>(string method, T @params)
-            {
-                Assert.Equal(RazorLanguageServerCustomMessageTargets.RazorDefinitionEndpointName, method);
-                var definitionParams = Assert.IsType<DelegatedPositionParams>(@params);
-
-                var definitionRequest = new TextDocumentPositionParams()
-                {
-                    TextDocument = new TextDocumentIdentifier()
-                    {
-                        Uri = _csharpDocumentUri
-                    },
-                    Position = definitionParams.ProjectedPosition
-                };
-
-                var result = await _csharpServer.ExecuteRequestAsync<TextDocumentPositionParams, DefinitionResult?>(Methods.TextDocumentDefinitionName, definitionRequest, CancellationToken.None);
-
-                return new TestResponseRouterReturn(result);
-            }
         }
     }
 }
